+++ src/upnp/xml_builder.h
@@ -111,19 +111,27 @@ public:
         const std::string& clientGroup,
         const std::map<std::string, std::string>& mimeMappings) const;
 
-    std::optional<std::string> renderContainerImageURL(const std::shared_ptr<CdsContainer>& cont) const;
-    std::optional<std::string> renderItemImageURL(const std::shared_ptr<CdsItem>& item) const;
+    /// \brief Renders the image of a container as url to be used in a resource
+    std::optional<std::string> renderContainerImageURL(
+        const std::shared_ptr<CdsContainer>& cont) const;
+    /// \brief Renders the image of an item as url to be used in a resource
+    std::optional<std::string> renderItemImageURL(
+        const std::shared_ptr<CdsItem>& item) const;
+    /// \brief Renders the subtilte of a video as url to be used in a resource
     std::optional<std::string> renderSubtitleURL(
         const std::shared_ptr<CdsItem>& item,
         const std::map<std::string, std::string>& mimeMappings) const;
 
+    /// \brief Renders a resource of an item
     std::string renderResourceURL(
         const CdsObject& item,
         const CdsResource& res,
         const std::map<std::string, std::string>& mimeMappings,
         const std::string& clientGroup = "") const;
 
-    void addResources(const std::shared_ptr<CdsItem>& item,
+    /// \brief Adds all resources to the output
+    void addResources(
+        const std::shared_ptr<CdsItem>& item,
         pugi::xml_node& parent,
         const std::vector<std::string>& filter,
         const std::shared_ptr<Quirks>& quirks) const;
@@ -133,7 +141,11 @@ public:
     std::string getFirstResourcePath(const std::shared_ptr<CdsItem>& item) const;
 
     /// \brief convert xml tree to string
-    static std::string printXml(const pugi::xml_node& entry, const char* indent = PUGIXML_TEXT("\t"), int flags = pugi::format_default);
+    static std::string printXml(
+        const pugi::xml_node& entry,
+        const char* indent = PUGIXML_TEXT("\t"),
+        int flags = pugi::format_default);
+    /// \brief make sure xml syntax elemnts are escaped correctly
     static std::string encodeEscapes(std::string s);
 
     std::string getDLNAContentHeader(
@@ -160,7 +172,19 @@ protected:
     std::map<std::string, std::string> containerPropertyDefaults;
     std::map<ConfigVal, std::map<std::string, std::string>> objectNamespaces;
 
+    /// \brief properties to tweak XML output
+    struct XmlStringFormat {
+        /// \brief string values have to be encoded
+        bool strictXml;
+        /// \brief only ASCII characters are allowed in outout
+        bool asciiXml;
+        /// \brief output strings have to be limited in length
+        std::size_t stringLimit;
+    };
+
+    /// \brief get resources in configured order
     std::deque<std::shared_ptr<CdsResource>> getOrderedResources(const CdsObject& object) const;
+    /// \brief insert pseudo resource for transcoding action
     std::pair<bool, int> insertTempTranscodingResource(
         const std::shared_ptr<CdsItem>& item,
         const std::shared_ptr<Quirks>& quirks,
@@ -182,8 +206,7 @@ protected:
         const std::string& key,
         const std::string& val) const;
     std::vector<std::string> addPropertyList(
-        bool strictXml,
-        std::size_t stringLimit,
+        const UpnpXMLBuilder::XmlStringFormat& xmlFormat,
         pugi::xml_node& result,
         const std::vector<std::string>& filter,
         const std::vector<std::pair<std::string, std::string>>& meta,
@@ -212,5 +235,9 @@ protected:
         const std::vector<std::string>& propNames,
         const std::vector<std::string>& filter,
         const std::map<std::string, std::string>& defaults);
+    ///  \brief update upnp property according to configuration
+    static std::string formatXmlString(
+        const UpnpXMLBuilder::XmlStringFormat& xmlFormat,
+        const std::string& input);
 };
 #endif // __UPNP_XML_H__