+++ ReleaseNotes.md
@@ -7,6 +7,7 @@ Sqlite backup is now enabled by default.
 ### NEW Features
 
 - Support "natural" file sorting for numbers in file names
+- Support transliteration of device output
 
 ### FIXES